# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Microclimate AI Weather Predictor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /locations:
    get:
      operationId: getLocations
      tags: [locations]
      summary: List tracked locations
      description: Returns all registered tracked locations (active and inactive)
      responses:
        "200":
          description: List of tracked locations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationsResponse"
    post:
      operationId: addLocation
      tags: [locations]
      summary: Add a tracked location
      description: Register a new location for automatic hourly weather collection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddLocationRequest"
      responses:
        "201":
          description: Location created
          content:
            application/json:
              schema:
                type: object
                properties:
                  location:
                    $ref: "#/components/schemas/TrackedLocation"
                required:
                  - location
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /locations/{id}/activate:
    put:
      operationId: activateLocation
      tags: [locations]
      summary: Activate a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Location activated
          content:
            application/json:
              schema:
                type: object
                properties:
                  location:
                    $ref: "#/components/schemas/TrackedLocation"
                required:
                  - location
        "404":
          description: Location not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /locations/{id}/deactivate:
    put:
      operationId: deactivateLocation
      tags: [locations]
      summary: Deactivate a location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Location deactivated
          content:
            application/json:
              schema:
                type: object
                properties:
                  location:
                    $ref: "#/components/schemas/TrackedLocation"
                required:
                  - location
        "404":
          description: Location not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /locations/{id}:
    delete:
      operationId: deleteLocation
      tags: [locations]
      summary: Delete a tracked location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Location deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: boolean
                  location:
                    $ref: "#/components/schemas/TrackedLocation"
                required:
                  - deleted
                  - location
        "404":
          description: Location not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    TrackedLocation:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - latitude
        - longitude
        - active
        - createdAt
    AddLocationRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
      required:
        - name
        - latitude
        - longitude
    LocationsResponse:
      type: object
      properties:
        locations:
          type: array
          items:
            $ref: "#/components/schemas/TrackedLocation"
      required:
        - locations
...