# lib/api-zod/src/generated/types/addLocationRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Microclimate AI Weather Predictor API
 * OpenAPI spec version: 0.1.0
 */

export interface AddLocationRequest {
  /**
   * @minLength 1
   * @maxLength 100
   */
  name: string;
  /**
   * @minimum -90
   * @maximum 90
   */
  latitude: number;
  /**
   * @minimum -180
   * @maximum 180
   */
  longitude: number;
}