# lib/api-zod/src/generated/types/trackedLocation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Microclimate AI Weather Predictor API
 * OpenAPI spec version: 0.1.0
 */

export interface TrackedLocation {
  id: number;
  name: string;
  latitude: number;
  longitude: number;
  active: boolean;
  createdAt: Date;
}